tools/requirements.txt
qingcloud-sdk
paramiko
